Added required fields to the upload page #68 #70
Added title, date, start time, and location fields to the upload page. The fields are required. If they aren't filled, a student should see warnings beneath the fields and shouldn't be able to upload. Hasn't been tested with the backend. If the student provides a file, may not prevent student from uploading.