Implemented basic upvote button functionality for posts #61
Added an UpvoteButton class to homepageCard.js and added it to the homepageCard. Every card starts with 2 upvotes. The upvotes are displayed on the button. Clicking the button once increments the count and fills in the button with the color blue. Clicking the button again will set upvotes back to the original upvote count and return the button to a blue outline (not filled in). Not sure if the count will stay updated after refreshing the page, or after moving to a different page then returning